style: fix clippy warnings

+++ crates/interledger-store-redis/tests/common/redis_helpers.rs
@@ -93,10 +93,7 @@ impl RedisServer {
         };
 
         let process = cmd.spawn().unwrap();
-        RedisServer {
-            process: process,
-            addr: addr,
-        }
+        RedisServer { process, addr }
     }
 
     pub fn wait(&mut self) {
@@ -110,11 +107,8 @@ impl RedisServer {
     pub fn stop(&mut self) {
         let _ = self.process.kill();
         let _ = self.process.wait();
-        match *self.get_client_addr() {
-            redis::ConnectionAddr::Unix(ref path) => {
-                fs::remove_file(&path).ok();
-            }
-            _ => {}
+        if let redis::ConnectionAddr::Unix(ref path) = *self.get_client_addr() {
+            fs::remove_file(&path).ok();
         }
     }
 }
@@ -160,10 +154,7 @@ impl TestContext {
         }
         redis::cmd("FLUSHDB").execute(&con);
 
-        TestContext {
-            server: server,
-            client: client,
-        }
+        TestContext { server, client }
     }
 
     // This one was added and not in the original file
